# Use per-task, per-device GPU streams for asynchronous execution