--- a/runtime/include/chpl-gpu-impl.h
+++ b/runtime/include/chpl-gpu-impl.h
@@ -35,17 +35,18 @@ void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
 void chpl_gpu_impl_launch_kernel_flat(int ln, int32_t fn,
                                  const char* name,
                                  int64_t num_threads, int blk_dim,
+                                 void* stream,
                                  int nargs, va_list args);
 
 void* chpl_gpu_impl_mem_alloc(size_t size);
-void* chpl_gpu_impl_mem_array_alloc(size_t size);
-void chpl_gpu_impl_mem_free(void* memAlloc);
-void* chpl_gpu_impl_memset(void* addr, const uint8_t val, size_t n);
+void* chpl_gpu_impl_mem_array_alloc(size_t size, void* stream);
+void chpl_gpu_impl_mem_free(void* memAlloc, void* stream);
+void* chpl_gpu_impl_memset(void* addr, const uint8_t val, size_t n, void* stream);
 void chpl_gpu_impl_hostmem_register(void *memAlloc, size_t size);
 
-void chpl_gpu_impl_copy_device_to_host(void* dst, const void* src, size_t n);
-void chpl_gpu_impl_copy_host_to_device(void* dst, const void* src, size_t n);
-void chpl_gpu_impl_copy_device_to_device(void* dst, const void* src, size_t n);
+void chpl_gpu_impl_copy_device_to_host(void* dst, const void* src, size_t n, void* stream);
+void chpl_gpu_impl_copy_host_to_device(void* dst, const void* src, size_t n, void* stream);
+void chpl_gpu_impl_copy_device_to_device(void* dst, const void* src, size_t n, void* stream);
 
 void* chpl_gpu_impl_comm_async(void *dst, void *src, size_t n);
 void chpl_gpu_impl_comm_wait(void *stream);
@@ -62,6 +63,9 @@ bool chpl_gpu_impl_can_access_peer(int dev1, int dev2);
 void chpl_gpu_impl_set_peer_access(int dev1, int dev2, bool enable);
 
 void chpl_gpu_impl_use_device(c_sublocid_t dev_id);
+
+void* chpl_gpu_impl_create_stream(void);
+void chpl_gpu_impl_destroy_stream(void* stream);
 
 #ifdef __cplusplus
 }