--- a/runtime/include/chpl-gpu-impl.h
+++ b/runtime/include/chpl-gpu-impl.h
@@ -45,9 +45,12 @@ void chpl_gpu_impl_mem_free(void* memAlloc, void* stream);
 void* chpl_gpu_impl_memset(void* addr, const uint8_t val, size_t n, void* stream);
 void chpl_gpu_impl_hostmem_register(void *memAlloc, size_t size);
 
-void chpl_gpu_impl_copy_device_to_host(void* dst, const void* src, size_t n, void* stream);
-void chpl_gpu_impl_copy_host_to_device(void* dst, const void* src, size_t n, void* stream);
-void chpl_gpu_impl_copy_device_to_device(void* dst, const void* src, size_t n, void* stream);
+void chpl_gpu_impl_copy_device_to_host(void* dst, const void* src, size_t n,
+                                       void* stream);
+void chpl_gpu_impl_copy_host_to_device(void* dst, const void* src, size_t n,
+                                       void* stream);
+void chpl_gpu_impl_copy_device_to_device(void* dst, const void* src, size_t n,
+                                         void* stream);
 
 void* chpl_gpu_impl_comm_async(void *dst, void *src, size_t n);
 void chpl_gpu_impl_comm_wait(void *stream);
@@ -65,6 +68,7 @@ void chpl_gpu_impl_set_peer_access(int dev1, int dev2, bool enable);
 
 void chpl_gpu_impl_use_device(c_sublocid_t dev_id);
 
+bool chpl_gpu_impl_supports_async_streams(int dev_id);
 void* chpl_gpu_impl_create_stream(void);
 void chpl_gpu_impl_destroy_stream(void* stream);
 void chpl_gpu_impl_stream_synchronize(void* stream);