--- a/runtime/include/chpl-gpu-impl.h
+++ b/runtime/include/chpl-gpu-impl.h
@@ -42,8 +42,9 @@ void chpl_gpu_impl_launch_kernel_flat(int ln, int32_t fn,
 void* chpl_gpu_impl_mem_alloc(size_t size);
 void* chpl_gpu_impl_mem_array_alloc(size_t size, void* stream);
 void chpl_gpu_impl_mem_free(void* memAlloc, void* stream);
-void* chpl_gpu_impl_memset(void* addr, const uint8_t val, size_t n, void* stream);
-void chpl_gpu_impl_hostmem_register(void *memAlloc, size_t size);
+void* chpl_gpu_impl_memset(void* addr, const uint8_t val, size_t n,
+                           void* stream);
+void chpl_gpu_impl_hostmem_register(void *memAlloc, size_t size);               
 
 void chpl_gpu_impl_copy_device_to_host(void* dst, const void* src, size_t n,
                                        void* stream);