--- a/runtime/include/chpl-gpu-impl.h
+++ b/runtime/include/chpl-gpu-impl.h
@@ -31,6 +31,7 @@ void chpl_gpu_impl_launch_kernel(int ln, int32_t fn,
                                  const char* name,
                                  int grd_dim_x, int grd_dim_y, int grd_dim_z,
                                  int blk_dim_x, int blk_dim_y, int blk_dim_z,
+                                 void* stream,
                                  int nargs, va_list args);
 void chpl_gpu_impl_launch_kernel_flat(int ln, int32_t fn,
                                  const char* name,