--- a/runtime/include/chpl-gpu-impl.h
+++ b/runtime/include/chpl-gpu-impl.h
@@ -44,7 +44,7 @@ void* chpl_gpu_impl_mem_array_alloc(size_t size, void* stream);
 void chpl_gpu_impl_mem_free(void* memAlloc, void* stream);
 void* chpl_gpu_impl_memset(void* addr, const uint8_t val, size_t n,
                            void* stream);
-void chpl_gpu_impl_hostmem_register(void *memAlloc, size_t size);               
+void chpl_gpu_impl_hostmem_register(void *memAlloc, size_t size);
 
 void chpl_gpu_impl_copy_device_to_host(void* dst, const void* src, size_t n,
                                        void* stream);
@@ -69,9 +69,11 @@ void chpl_gpu_impl_set_peer_access(int dev1, int dev2, bool enable);
 
 void chpl_gpu_impl_use_device(c_sublocid_t dev_id);
 
+// TODO this should return true for NVIDIA, too. The current NVIDIA
+// implementation is returning true/false only for allocators
 bool chpl_gpu_impl_supports_async_streams(int dev_id);
-void* chpl_gpu_impl_create_stream(void);
-void chpl_gpu_impl_destroy_stream(void* stream);
+void* chpl_gpu_impl_stream_create(void);
+void chpl_gpu_impl_stream_destroy(void* stream);
 void chpl_gpu_impl_stream_synchronize(void* stream);
 
 #ifdef __cplusplus