--- a/runtime/include/chpl-gpu-impl.h
+++ b/runtime/include/chpl-gpu-impl.h
@@ -69,6 +69,7 @@ void chpl_gpu_impl_set_peer_access(int dev1, int dev2, bool enable);
 
 void chpl_gpu_impl_use_device(c_sublocid_t dev_id);
 
+void chpl_gpu_impl_synchronize(void);
 bool chpl_gpu_impl_stream_supported(void);
 void* chpl_gpu_impl_stream_create(void);
 void chpl_gpu_impl_stream_destroy(void* stream);