--- a/runtime/include/chpl-gpu.h
+++ b/runtime/include/chpl-gpu.h
@@ -35,6 +35,7 @@ extern "C" {
 extern bool chpl_gpu_debug;
 extern int chpl_gpu_num_devices;
 extern bool chpl_gpu_no_cpu_mode_warning;
+extern bool chpl_gpu_always_sync_kernels;
 
 
 #ifdef HAS_GPU_LOCALE