--- a/runtime/include/chpl-gpu.h
+++ b/runtime/include/chpl-gpu.h
@@ -53,7 +53,7 @@ static inline void CHPL_GPU_DEBUG(const char *str, ...) {
 
 #ifdef CHPL_GPU_ENABLE_PROFILE
 // returns time from epoch in milliseconds. Used in macros below.
-static inline long double get_time() {
+static inline long double get_time(void) {
   struct timeval tv;
 
   gettimeofday(&tv, NULL);