--- a/runtime/include/chpl-mem-consistency.h
+++ b/runtime/include/chpl-mem-consistency.h
@@ -24,7 +24,7 @@
 #include "chpl-atomics.h" // for memory_order
 
 #include "chpl-cache.h" // for chpl_cache_release, chpl_cache_acquire
-			//
+
 #include "chpl-gpu.h" // for chpl_gpu_task_fence
 
 #ifdef __cplusplus
@@ -81,11 +81,6 @@ void chpl_rmem_consist_acquire(int ln, int32_t fn)
 #ifdef HAS_CHPL_CACHE_FNS
   chpl_cache_acquire(ln, fn);
 #endif
-  /*
-#ifdef HAS_GPU_LOCALE
-  chpl_gpu_task_fence();
-#endif
-*/
 }