--- a/test/gpu/native/asynchrony/asyncTaskComm.execopts
+++ b/test/gpu/native/asynchrony/asyncTaskComm.execopts
@@ -1,0 +1,2 @@
+--printSum
+--printSum --multitask