… runtime/src/gpu/amd/gpu-amd.c
… runtime/src/gpu/cpu/gpu-cpu.c
… runtime/src/gpu/nvidia/gpu-nvidia.c
… test/gpu/native/asynchrony/asyncTaskComm.chpl
… test/gpu/native/asynchrony/asyncTaskComm.good
… test/gpu/native/asynchrony/syncFlag-forcedSync.good
… test/gpu/native/asynchrony/syncFlag-noForcedSync.good
… test/gpu/native/asynchrony/syncFlag.chpl
… test/gpu/native/asynchrony/syncFlag.compopts
… test/gpu/native/asynchrony/syncFlag.execopts
… test/gpu/native/asynchrony/syncFlag.good
… test/gpu/native/asynchrony/syncFlag.prediff
… test/gpu/native/asynchrony/syncFlag.skipif
… test/gpu/native/diags.prediff