@e-kayrakli 2023-09-11
@stonea / @ronawho -- I requested reviews from both of you for the GPU and the higher-level runtime changes, respectively. I am in the process of adding/fixing some new tests, but I think this is ready for review. I'll try to post some performance data under here shortly, as well.