@e-kayrakli 2023-09-11
Running `asyncTaskComm` added in this PR with 200 reps and `real`s (i.e. total data size is 16G):

Data Size   In Elems | Per Chunk Size in Elems | Number of Workers | `main` Throughput (GB/s) | branch Throughput (GB/s) | Speedup over Optimal Main (x)
-- | -- | -- | -- | -- | --
1B | Single Task | N/A | 5.85 | 5.77 | 0.99
1B | 1M | 2 | 4.63 | 6.71 | 1.15
1B | 1M | 4 | 4.57 | 6.74 | 1.15
1B | 1M | 8 | 4.58 | 6.66 | 1.14
1B | 1M | 16 | 4.55 | 6.64 | 1.14
1B | 10M | 2 | 5.53 | 8.14 | 1.39
1B | 10M | 4 | 5.50 | 8.33 | 1.42
1B | 10M | 8 | 4.88 | 8.17 | 1.40
1B | 10M | 16 | 5.40 | 7.97 | 1.36
1B | 100M | 2 | 5.70 | 8.41 | 1.44
1B | 100M | 4 | 5.61 | 8.05 | 1.38
1B | 100M | 8 | 5.49 | 7.77 | 1.33
1B | 100M | 16 | 4.70 | 7.27 | 1.24



1. `main` throughput is always lower when you try to multitask
2. `branch` throughput is always higher when you try to multitask
3. Top speedup is 1.44x


---

Repeating the similar experiment with even larger data (total data size is 160G, can't fit in GPU memory)



Data Size   In Elems | Per Chunk Size in Elems | Number of Workers | `main` Throughput (GB/s) | branch Throughput (GB/s)
-- | -- | -- | -- | --
10B | Single Task | N/A | OOM | OOM
10B | 1M | 2 | OOM | 6.79
10B | 1M | 4 | OOM | 6.86
10B | 1M | 8 | OOM | 6.81
10B | 1M | 16 | OOM | 6.80
10B | 10M | 2 | OOM | 8.51
10B | 10M | 4 | OOM | 8.61
10B | 10M | 8 | OOM | 8.50
10B | 10M | 16 | OOM | 8.47
10B | 100M | 2 | OOM | 8.68
10B | 100M | 4 | OOM | 8.75
10B | 100M | 8 | OOM | 8.65
10B | 100M | 16 | OOM | 8.50
10B | 1B | 2 | OOM | 8.68
10B | 1B | 4 | OOM | OOM
10B | 1B | 8 | OOM | OOM
10B | 1B | 16 | OOM | OOM

The branch is able to achieve similar/better throughput as before with larger data.