@e-kayrakli 2023-09-12
@stonea / @ronawho while I will work on adding tests + supporting the cpu-as-device mode + making sure CI pass, I think I've addressed the major concerns from the meeting earlier today. They were

- we should be synchronizing within a stream all the time: https://github.com/chapel-lang/chapel/pull/23307/commits/60e2540af69fca38f575778cb754206965e23c2d does that. I don't observe any performance hit from this.
- yielding while waiting for sync on a stream: this actually hurts performance in the artificial benchmark I have. We might want that support down the road though. So I have the implementation that's hardwired to be off. See https://github.com/chapel-lang/chapel/pull/23307/commits/8f419188cabcbb1e7632be9a4efdc02fafefdecb
- Cost of stream creation: On my workstation, I see about 8-9 microseconds for stream creation including some memory allocation for the stream array. Subsequent accesses to streams are not observable performance wise. On my system, I am getting around 25 microsecond for empty kernel launch for comparison. I am not worried at all by these numbers. https://github.com/chapel-lang/chapel/pull/23307/commits/47b783e4c30d897334cf39b04645a2fe7d8569d8 adds internal profiling support for that.
- `unified_memory` should always be synchronized: I got a bit worried that there can be inter-device communication that can't be captured if two devices use independent streams. So https://github.com/chapel-lang/chapel/pull/23307/commits/19e9b1a46d0b4787432fb3c11f75ab689694bde9 makes `unified_memory` use the default stream always. This is essentially the same behavior we have on `main` today. Note also that even with independent streams `unified_memory` performance is 2x as bad as the default `array_on_device` performance.

I think this is ready for a review.